Update events and store array of stats
Events sent will be broken down into smaller chunks, so entire lists
are not resent, thus saving bandwidth and increasing efficiency. In
stats-page, a function will be setup so that the most recent stats
are always available as a binding.

The full list of stats was decided to be skept so that graphs
containing statistics from the past 6 hours will be available on
stats-page.