Add chart to stats-page and add dummy stat generation
A chart will be added to stats-page, using Google Charts. Styling
will have been done, and the appropriate event listeners trigger
a chart update. A possible ficx for the future may be to make
updateChart a private method and somehow detect when the set method
is called from outside the element.

Also, every 2 minutes a stat update will be issued if there are no
clients are connected, making the data consistent for the chart.