Migrate from Google Charts to Highcharts
Google Charts was far too slow to load with large amounts of data.
Switching to Highcharts seems to have fixed that issue, and increased
the overall reposiveness of the site, especially upon page load.

The chart will still need to be made to fit it's container after the
element size has updated. This may be fixed by changing the ready
event listener to an attached listener.