Fix chart reflowing and minor improvements
The chart reflows at a 100ms interval from the time stats-page is
attached to the DOM. This means it will will the area it has been
designated.

Move the unit of speed (K/sec) into the axis title, because on
mobile it takes up too much space horizontally to show it on the
labels.

Also, remove unnecessary logging.