Add icons to header and correct element alignment
Save icon-white.svg as a plain SVG (rather than an Inkscape one) so it is more
compressed and can be used with iron-iconset-svg. Then, copy the relevant parts
from icon-white.svg into site-icons.html and add an iron-icon into the header to
display the icon. Also, add paper-icon-button for dropdown menu.